requirements.txt
aniso8601==7.0.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-ckeditor==5.7.1
django-cors-headers==3.2.0
django-easy-select2==1.5.6
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-graphql-social-auth==0.1.4
django-js-asset==1.2.2
graphene==2.1.8
graphene-django==2.6.0
graphene-file-upload==1.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
pkg-resources==0.0.0
promise==2.2.1
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
opencv-python-headless=4.1.2.30
